+++ server/historian/packages/historian-base/src/routes/v1/git/blobs.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { ICommit, ICreateCommitParams } from "@fluidframework/gitresources";
+import * as git from "@fluidframework/gitresources";
 import {
 	IStorageNameRetriever,
 	IThrottler,
@@ -15,9 +15,14 @@ import { validateRequestParams } from "@fluidframework/server-services-shared";
 import { Router } from "express";
 import * as nconf from "nconf";
 import winston from "winston";
-import { ICache, IDenyList, ITenantService, ISimplifiedCustomDataRetriever } from "../../services";
-import * as utils from "../utils";
-import { Constants } from "../../utils";
+import {
+	ICache,
+	IDenyList,
+	ITenantService,
+	ISimplifiedCustomDataRetriever,
+} from "../../../services";
+import * as utils from "../../utils";
+import { Constants } from "../../../utils";
 
 export function create(
 	config: nconf.Provider,
@@ -42,11 +47,11 @@ export function create(
 		Constants.generalRestCallThrottleIdPrefix,
 	);
 
-	async function createCommit(
+	async function createBlob(
 		tenantId: string,
 		authorization: string | undefined,
-		params: ICreateCommitParams,
-	): Promise<ICommit> {
+		body: git.ICreateBlobParams,
+	): Promise<git.ICreateBlobResponse> {
 		const service = await utils.createGitService({
 			config,
 			tenantId,
@@ -59,15 +64,15 @@ export function create(
 			ephemeralDocumentTTLSec,
 			simplifiedCustomDataRetriever,
 		});
-		return service.createCommit(params);
+		return service.createBlob(body);
 	}
 
-	async function getCommit(
+	async function getBlob(
 		tenantId: string,
 		authorization: string | undefined,
 		sha: string,
 		useCache: boolean,
-	): Promise<ICommit> {
+	): Promise<git.IBlob> {
 		const service = await utils.createGitService({
 			config,
 			tenantId,
@@ -79,40 +84,96 @@ export function create(
 			denyList,
 			ephemeralDocumentTTLSec,
 		});
-		return service.getCommit(sha, useCache);
+		return service.getBlob(sha, useCache);
 	}
 
+	/**
+	 * Historian https ping endpoint for availability monitoring system
+	 */
+	router.get(
+		"/repos/ping",
+		throttle(restTenantGeneralThrottler, winston, {
+			...tenantThrottleOptions,
+			throttleIdPrefix: "ping",
+		}),
+		// eslint-disable-next-line @typescript-eslint/no-misused-promises
+		async (request, response) => {
+			response.sendStatus(200);
+		},
+	);
+
 	router.post(
-		"/repos/:ignored?/:tenantId/git/commits",
+		"/repos/:ignored?/:tenantId/git/blobs",
 		validateRequestParams("tenantId"),
 		throttle(restTenantGeneralThrottler, winston, tenantThrottleOptions),
 		utils.verifyToken(revokedTokenChecker),
 		(request, response, next) => {
-			const commitP = createCommit(
+			const blobP = createBlob(
 				request.params.tenantId,
 				request.get("Authorization"),
 				request.body,
 			);
-
-			utils.handleResponse(commitP, response, false, undefined, 201);
+			utils.handleResponse(blobP, response, false, 201);
 		},
 	);
 
+	/**
+	 * Retrieves the given blob from the repository
+	 */
 	router.get(
-		"/repos/:ignored?/:tenantId/git/commits/:sha",
+		"/repos/:ignored?/:tenantId/git/blobs/:sha",
 		validateRequestParams("tenantId", "sha"),
 		throttle(restTenantGeneralThrottler, winston, tenantThrottleOptions),
 		utils.verifyToken(revokedTokenChecker),
 		(request, response, next) => {
 			const useCache = !("disableCache" in request.query);
-			const commitP = getCommit(
+			const blobP = getBlob(
+				request.params.tenantId,
+				request.get("Authorization"),
+				request.params.sha,
+				useCache,
+			);
+			utils.handleResponse(blobP, response, useCache);
+		},
+	);
+
+	/**
+	 * Retrieves the given blob as an image
+	 */
+	router.get(
+		"/repos/:ignored?/:tenantId/git/blobs/raw/:sha",
+		validateRequestParams("tenantId", "sha"),
+		throttle(restTenantGeneralThrottler, winston, tenantThrottleOptions),
+		utils.verifyToken(revokedTokenChecker),
+		(request, response, next) => {
+			const useCache = !("disableCache" in request.query);
+
+			const blobP = getBlob(
 				request.params.tenantId,
 				request.get("Authorization"),
 				request.params.sha,
 				useCache,
 			);
 
-			utils.handleResponse(commitP, response, useCache);
+			blobP
+				.then((blob) => {
+					if (useCache) {
+						response.setHeader("Cache-Control", "public, max-age=31536000");
+					}
+					// Make sure the browser will expose specific headers for performance analysis.
+					response.setHeader(
+						"Access-Control-Expose-Headers",
+						"Content-Encoding, Content-Length, Content-Type",
+					);
+					// In order to report W3C timings, Time-Allow-Origin needs to be set.
+					response.setHeader("Timing-Allow-Origin", "*");
+					response
+						.status(200)
+						.write(Buffer.from(blob.content, "base64"), () => response.end());
+				})
+				.catch((error) => {
+					response.status(error?.code ?? 400).json(error?.message ?? error);
+				});
 		},
 	);
