+++ server/historian/packages/historian-base/src/routes/v1/repository/headers.ts
@@ -3,6 +3,7 @@
  * Licensed under the MIT License.
  */
 
+import { IHeader } from "@fluidframework/gitresources";
 import {
 	IStorageNameRetriever,
 	IThrottler,
@@ -14,9 +15,14 @@ import { validateRequestParams } from "@fluidframework/server-services-shared";
 import { Router } from "express";
 import * as nconf from "nconf";
 import winston from "winston";
-import { ICache, IDenyList, ITenantService, ISimplifiedCustomDataRetriever } from "../../services";
-import * as utils from "../utils";
-import { Constants } from "../../utils";
+import {
+	ICache,
+	IDenyList,
+	ITenantService,
+	ISimplifiedCustomDataRetriever,
+} from "../../../services";
+import * as utils from "../../utils";
+import { Constants } from "../../../utils";
 
 export function create(
 	config: nconf.Provider,
@@ -41,11 +47,31 @@ export function create(
 		Constants.generalRestCallThrottleIdPrefix,
 	);
 
-	async function getContent(
+	async function getHeader(
 		tenantId: string,
 		authorization: string | undefined,
-		path: string,
-		ref: string | undefined,
+		sha: string,
+		useCache: boolean,
+	): Promise<IHeader> {
+		const service = await utils.createGitService({
+			config,
+			tenantId,
+			authorization,
+			tenantService,
+			storageNameRetriever,
+			documentManager,
+			cache,
+			denyList,
+			ephemeralDocumentTTLSec,
+		});
+		return service.getHeader(sha, useCache);
+	}
+
+	async function getTree(
+		tenantId: string,
+		authorization: string | undefined,
+		sha: string,
+		useCache: boolean,
 	): Promise<any> {
 		const service = await utils.createGitService({
 			config,
@@ -58,22 +84,40 @@ export function create(
 			denyList,
 			ephemeralDocumentTTLSec,
 		});
-		return service.getContent(path, ref);
+		return service.getFullTree(sha, useCache);
 	}
 
 	router.get(
-		"/repos/:ignored?/:tenantId/contents/*",
-		validateRequestParams("tenantId", 0),
+		"/repos/:ignored?/:tenantId/headers/:sha",
+		validateRequestParams("tenantId", "sha"),
 		throttle(restTenantGeneralThrottler, winston, tenantThrottleOptions),
 		utils.verifyToken(revokedTokenChecker),
 		(request, response, next) => {
-			const contentP = getContent(
+			const useCache = !("disableCache" in request.query);
+			const headerP = getHeader(
 				request.params.tenantId,
 				request.get("Authorization"),
-				request.params[0],
-				utils.queryParamToString(request.query.ref),
+				request.params.sha,
+				useCache,
 			);
-			utils.handleResponse(contentP, response, false);
+			utils.handleResponse(headerP, response, useCache);
+		},
+	);
+
+	router.get(
+		"/repos/:ignored?/:tenantId/tree/:sha",
+		validateRequestParams("tenantId", "sha"),
+		throttle(restTenantGeneralThrottler, winston, tenantThrottleOptions),
+		utils.verifyToken(revokedTokenChecker),
+		(request, response, next) => {
+			const useCache = !("disableCache" in request.query);
+			const headerP = getTree(
+				request.params.tenantId,
+				request.get("Authorization"),
+				request.params.sha,
+				useCache,
+			);
+			utils.handleResponse(headerP, response, useCache);
 		},
 	);
