+++ server/historian/packages/historian-base/src/routes/v1/summaries.ts
@@ -20,9 +20,9 @@ import { Router } from "express";
 import * as nconf from "nconf";
 import winston from "winston";
 import { BaseTelemetryProperties, Lumberjack } from "@fluidframework/server-services-telemetry";
-import { ICache, IDenyList, ITenantService, ISimplifiedCustomDataRetriever } from "../services";
-import { parseToken, Constants } from "../utils";
-import * as utils from "./utils";
+import { ICache, IDenyList, ITenantService, ISimplifiedCustomDataRetriever } from "../../services";
+import { parseToken, Constants } from "../../utils";
+import * as utils from "../utils";
 
 export function create(
 	config: nconf.Provider,
